change number of bands to be stored as int instead of size_t, emit numBandsROIChanged in ImageModel, default nbands handling in ImageModel::spawn()
-1 is used as a special value (uninitialized) for the number of bands. Previously the number of bands was stored in a size_t. This lead to the comparison "bands > getNumBandsFull()" being true always. This problem is fixed now by using int everywhere. As a side note: gcc does _not_ warn on assigning a negative literal to a variable of unsigned type.

+++ gerbil/gui/mainwindow.cpp
@@ -123,8 +123,8 @@ void MainWindow::initSignals(Controller *chief)
 			this, SLOT(bandsSliderMoved(int)));
 	connect(bandsSlider, SIGNAL(sliderMoved(int)),
 			this, SLOT(bandsSliderMoved(int)));
-	connect(this, SIGNAL(specRescaleRequested(size_t)),
-			chief, SLOT(rescaleSpectrum(size_t)));
+	connect(this, SIGNAL(specRescaleRequested(int)),
+			chief, SLOT(rescaleSpectrum(int)));
 
 	/// global shortcuts
 	QShortcut *scr = new QShortcut(Qt::CTRL + Qt::Key_S, this);
