+++ gerbil/gui/mainwindow.h
@@ -84,7 +84,7 @@ public slots:
 signals:
 	void ignoreLabelsRequested(bool);
 	void singleLabelRequested(bool);
-	void specRescaleRequested(size_t bands);
+	void specRescaleRequested(int bands);
 	void clearLabelRequested(short index);
 	void alterLabelRequested(short index, const cv::Mat1b &mask, bool negative);
